created examples module
