+++ iwp/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "iwp"
+name = "imagepkg"
 version = "0.1.0"
 edition = "2021"
