+++ iwp/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod examples;
 pub mod format;
 pub mod img;
 pub mod iwp;
