+++ iwp/src/mr.rs
@@ -108,191 +108,13 @@ pub fn morph_reconstruction(
 }
 
 mod tests {
-    use image::{GrayImage, ImageBuffer};
 
-    use crate::format::*;
+    #![allow(unused_imports)]
+
+    use crate::examples::*;
+    use crate::format;
     use crate::iwp;
     use crate::mr::*;
-
-    fn _gen_same_value_image(width: u32, height: u32, value: u8) -> ImageBuffer<Luma<u8>, Vec<u8>> {
-        let mut img = GrayImage::new(width, height);
-        for i in 0..width {
-            for j in 0..height {
-                img.put_pixel(i, j, Luma([value]))
-            }
-        }
-
-        return img;
-    }
-
-    /*
-    Gens the 6 x 6 image below
-    0 0 0 0 0 0
-    0 1 1 0 0 0
-    0 1 1 0 0 0
-    0 0 0 1 1 0
-    0 0 0 1 1 0
-    0 0 0 0 0 0
-    */
-    fn _gen_example_img() -> ImageBuffer<Luma<u8>, Vec<u8>> {
-        let mut img = _gen_same_value_image(6, 6, 0);
-
-        img.put_pixel(1, 1, Luma([1]));
-        img.put_pixel(1, 2, Luma([1]));
-        img.put_pixel(2, 1, Luma([1]));
-        img.put_pixel(2, 2, Luma([1]));
-        img.put_pixel(3, 3, Luma([1]));
-        img.put_pixel(3, 4, Luma([1]));
-        img.put_pixel(4, 3, Luma([1]));
-        img.put_pixel(4, 4, Luma([1]));
-
-        return img;
-    }
-
-    /*
-    Gens the 10 x 10 image below
-    08 08 08 08 08 08 08 08 08 08
-    08 12 12 12 08 08 09 08 09 08
-    08 12 12 12 08 08 08 09 08 08
-    08 12 12 12 08 08 09 08 09 08
-    08 08 08 08 08 08 08 08 08 08
-    08 09 08 08 08 16 16 16 08 08
-    08 08 08 09 08 16 16 16 08 08
-    08 08 09 08 08 16 16 16 08 08
-    08 09 08 09 08 08 08 08 08 08
-    08 08 08 08 08 08 09 08 08 08
-    */
-    fn _gen_big_marker_img() -> ImageBuffer<Luma<u8>, Vec<u8>> {
-        let mut base_img = _gen_same_value_image(10, 10, 8);
-
-        for i in 1..4 {
-            for j in 1..4 {
-                base_img.put_pixel(i, j, Luma([12]));
-            }
-        }
-
-        for i in 5..8 {
-            for j in 5..8 {
-                base_img.put_pixel(i, j, Luma([16]));
-            }
-        }
-
-        base_img.put_pixel(1, 5, Luma([9]));
-        base_img.put_pixel(1, 8, Luma([9]));
-        base_img.put_pixel(2, 7, Luma([9]));
-        base_img.put_pixel(3, 6, Luma([9]));
-        base_img.put_pixel(3, 8, Luma([9]));
-        base_img.put_pixel(6, 1, Luma([9]));
-        base_img.put_pixel(6, 3, Luma([9]));
-        base_img.put_pixel(6, 9, Luma([9]));
-        base_img.put_pixel(7, 2, Luma([9]));
-        base_img.put_pixel(8, 1, Luma([9]));
-        base_img.put_pixel(8, 3, Luma([9]));
-
-        return base_img;
-    }
-
-    /*
-    Gens the 10 x 10 image below
-    10 10 10 10 10 10 10 10 10 10
-    10 14 14 14 10 10 11 10 11 10
-    10 14 14 14 10 10 10 11 10 10
-    10 14 14 14 10 10 11 10 11 10
-    10 10 10 10 10 10 10 10 10 10
-    10 11 10 10 10 16 16 16 10 10
-    10 10 10 11 10 16 16 16 10 10
-    10 10 11 10 10 16 16 16 10 10
-    10 11 10 11 10 10 10 10 10 10
-    10 10 10 10 10 10 11 10 10 10
-    */
-    fn _gen_big_mask_img() -> ImageBuffer<Luma<u8>, Vec<u8>> {
-        let mut base_img = _gen_same_value_image(10, 10, 10);
-
-        for i in 1..4 {
-            for j in 1..4 {
-                base_img.put_pixel(i, j, Luma([14]));
-            }
-        }
-
-        for i in 5..8 {
-            for j in 5..8 {
-                base_img.put_pixel(i, j, Luma([18]));
-            }
-        }
-
-        base_img.put_pixel(1, 5, Luma([11]));
-        base_img.put_pixel(1, 8, Luma([11]));
-        base_img.put_pixel(2, 7, Luma([11]));
-        base_img.put_pixel(3, 6, Luma([11]));
-        base_img.put_pixel(3, 8, Luma([11]));
-        base_img.put_pixel(6, 1, Luma([11]));
-        base_img.put_pixel(6, 3, Luma([11]));
-        base_img.put_pixel(6, 9, Luma([11]));
-        base_img.put_pixel(7, 2, Luma([11]));
-        base_img.put_pixel(8, 1, Luma([11]));
-        base_img.put_pixel(8, 3, Luma([11]));
-
-        return base_img;
-    }
-
-    /*
-    Gens the 10 x 10 image below
-    10 10 10 10 10 10 10 10 10 10
-    10 12 12 12 10 10 10 10 10 10
-    10 12 12 12 10 10 10 10 10 10
-    10 12 12 12 10 10 10 10 10 10
-    10 10 10 10 10 10 10 10 10 10
-    10 10 10 10 10 16 16 16 10 10
-    10 10 10 10 10 16 16 16 10 10
-    10 10 10 10 10 16 16 16 10 10
-    10 10 10 10 10 10 10 10 10 10
-    10 10 10 10 10 10 10 10 10 10
-    */
-    fn _gen_expected_img() -> ImageBuffer<Luma<u8>, Vec<u8>> {
-        let mut base_img = _gen_same_value_image(10, 10, 10);
-
-        for i in 1..4 {
-            for j in 1..4 {
-                base_img.put_pixel(i, j, Luma([12]));
-            }
-        }
-
-        for i in 5..8 {
-            for j in 5..8 {
-                base_img.put_pixel(i, j, Luma([16]));
-            }
-        }
-
-        return base_img;
-    }
-
-    #[test]
-    fn test_get_pixel_neighbours() {
-        let mask = _gen_example_img();
-        let ngbs = img::get_pixel_neighbours(&mask, (0, 0), img::ConnTypes::Eight);
-        let expected = vec![(0, 1), (1, 0), (1, 1)];
-
-        assert_eq!(ngbs, expected);
-
-        let ngbs = img::get_pixel_neighbours(&mask, (0, 0), img::ConnTypes::Four);
-        let expected = vec![(0, 1), (1, 0)];
-
-        assert_eq!(ngbs, expected);
-
-        let ngbs = img::get_pixel_neighbours(&mask, (2, 2), img::ConnTypes::Eight);
-        let expected = vec![
-            (1, 1),
-            (1, 2),
-            (1, 3),
-            (2, 1),
-            (2, 3),
-            (3, 1),
-            (3, 2),
-            (3, 3),
-        ];
-
-        assert_eq!(ngbs, expected);
-    }
 
     #[test]
     fn test_get_initial_pixels() {
